2025-11-11 배포2

See merge request s13-final/S13P31A406!107

+++ backend/src/main/java/com/a404/duckonback/controller/MemeController.java
@@ -99,8 +99,7 @@ public class MemeController {
             @AuthenticationPrincipal CustomUserPrincipal principal,
             @RequestBody MemeUsageLogRequestDTO request
     ) {
-        Long userId = principal.getId();
-        memeUsageLogService.logMemeUsage(userId, request.getMemeId(), request.getUsageType());
+        memeUsageLogService.logMemeUsage(principal, request.getMemeId(), request.getUsageType());
 
         return ResponseEntity.ok(ApiResponseDTO.success(SuccessCode.MEME_USAGE_LOG_SUCCESS, null));
     }
@@ -111,6 +110,8 @@ public class MemeController {
             @AuthenticationPrincipal CustomUserPrincipal principal,
             @PathVariable Long memeId
     ) {
+        if(principal == null) throw new CustomException("로그인이 필요합니다.", ErrorCode.USER_NOT_AUTHENTICATED);
+
         memeService.createFavorite(principal.getId(), memeId);
         return ResponseEntity.ok(ApiResponseDTO.success(SuccessCode.MEME_FAVORITE_CREATED, null));
     }
@@ -121,6 +122,8 @@ public class MemeController {
             @AuthenticationPrincipal CustomUserPrincipal principal,
             @PathVariable Long memeId
     ) {
+        if(principal == null) throw new CustomException("로그인이 필요합니다.", ErrorCode.USER_NOT_AUTHENTICATED);
+
         memeService.deleteFavorite(principal.getId(), memeId);
         return ResponseEntity.ok(ApiResponseDTO.success(SuccessCode.MEME_FAVORITE_DELETED, null));
     }
@@ -162,6 +165,18 @@ public class MemeController {
         return ResponseEntity.ok(
                 ApiResponseDTO.success(SuccessCode.MEME_TOP10_RETRIEVE_SUCCESS, res)
         );
+    }
+
+    @Operation(
+            summary = "밈 상세 조회",
+            description = "밈 기본정보, 태그 이름 리스트, 생성자 정보를 반환합니다."
+    )
+    @GetMapping("/{memeId}/detail")
+    public ResponseEntity<ApiResponseDTO<MemeDetailDTO>> getMemeDetail(
+            @PathVariable Long memeId
+    ) {
+        MemeDetailDTO detail = memeService.getMemeDetail(memeId);
+        return ResponseEntity.ok(ApiResponseDTO.success(SuccessCode.MEME_RETRIEVE_SUCCESS, detail));
     }
 
     @Operation(
